Fix TIR truncation bug and use design system colors
- Fix TIR comparison: use Double range instead of truncated Int range
  (matches StatsCalculator pattern, prevents incorrect classification
  at fractional thresholds)
- Add TirGood and TintGood to Color.kt design system
- Use TirGood/TintGood in MealStatsTab instead of hardcoded hex values

+++ app/src/main/java/com/psjostrom/strimma/data/meal/MealAnalyzer.kt
@@ -176,7 +176,7 @@ class MealAnalyzer @Inject constructor() {
     ): Double {
         if (readings.isEmpty()) return 0.0
 
-        val inRange = readings.count { it.sgv in bgLow.toInt()..bgHigh.toInt() }
+        val inRange = readings.count { it.sgv.toDouble() in bgLow..bgHigh }
         return (inRange.toDouble() / readings.size) * PERCENT
     }
